requirements.txt
django
pillow
gunicorn
whitenoise
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
pip==24.3.1
sqlparse==0.5.3
tzdata==2025.1
gunicorn
pillow
whitenoise
django